+++ app/src/main/java/com/example/ottoplay/ClassDiagrams/DynamicWaypoint.java
@@ -1,9 +1,9 @@
-package com.example.ottoplay;
+package com.example.ottoplay.ClassDiagrams;
 
 public class DynamicWaypoint extends Waypoint {
     private String lastUpdateTime;
 
-    DynamicWaypoint(int wpId, int ownerId, String wpName) {
+    public DynamicWaypoint(int wpId, int ownerId, String wpName) {
         this.globalId = wpId;
         this.ownerUserId = ownerId;
         this.waypointName = wpName;
