+++ app/src/main/res/layout/activity_main.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".WaypointMapTestingSetupActivity">
+    tools:context=".WaypointMap.WaypointMapTestingSetupActivity">
 
     <TextView
         android:layout_width="wrap_content"
