+++ app/src/main/res/layout/activity_waypoint_map.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".WaypointMapActivity">
+    tools:context=".WaypointMap.WaypointMapActivity">
 
     <fragment
         android:id="@+id/g_map"
