+++ app/src/main/res/layout/activity_waypoint_playlists_from_map.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".WaypointPlaylistsFromMapActivity">
+    tools:context=".WaypointMap.WaypointPlaylistsFromMapActivity">
 
     <TextView
         android:id="@+id/title"
