+++ app/src/main/res/layout/waypointsettings.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".waypointSettingsActivity">
+    tools:context=".WaypointMap.waypointSettingsActivity">
 
     <LinearLayout
         android:id="@+id/mainHorizontalLayout"
